// webapp/src/lib/api/model/accountType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Kakeibo API
 * マルチテナント家計簿システム API
 * OpenAPI spec version: 1.0.0
 */

export type AccountType = (typeof AccountType)[keyof typeof AccountType];

export const AccountType = {
  cash: "cash",
  bank: "bank",
  credit_card: "credit_card",
  e_money: "e_money",
  investment: "investment",
} as const;